feature: Add Bootstrap validation to Login page form
